+++ src/Models/AttachmentFormat.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 /**
  * @property string $attachment_id
  * @property string $format
+ * @property \Carbon\Carbon $updated_at
  */
 class AttachmentFormat extends Model
 {